--- mpc_environment.yaml
@@ -1,48 +1,0 @@
-name: mpc
-channels:
-  - defaults
-  - https://repo.anaconda.com/pkgs/main
-  - https://repo.anaconda.com/pkgs/r
-dependencies:
-  - _libgcc_mutex=0.1=main
-  - _openmp_mutex=5.1=1_gnu
-  - bzip2=1.0.8=h5eee18b_6
-  - ca-certificates=2025.2.25=h06a4308_0
-  - ld_impl_linux-64=2.40=h12ee557_0
-  - libffi=3.4.4=h6a678d5_1
-  - libgcc-ng=11.2.0=h1234567_1
-  - libgomp=11.2.0=h1234567_1
-  - libstdcxx-ng=11.2.0=h1234567_1
-  - libuuid=1.41.5=h5eee18b_0
-  - ncurses=6.4=h6a678d5_0
-  - openssl=3.0.15=h5eee18b_0
-  - pip=25.0=py310h06a4308_0
-  - python=3.10.16=he870216_1
-  - readline=8.2=h5eee18b_0
-  - setuptools=75.8.0=py310h06a4308_0
-  - sqlite=3.45.3=h5eee18b_0
-  - tk=8.6.14=h39e8969_0
-  - tzdata=2025a=h04d1e81_0
-  - wheel=0.45.1=py310h06a4308_0
-  - xz=5.6.4=h5eee18b_1
-  - yaml=0.2.5=h7b6447c_0
-  - zlib=1.2.13=h5eee18b_1
-  - pip:
-      - casadi==3.6.7
-      - catkin-pkg==1.0.0
-      - contourpy==1.3.1
-      - cycler==0.12.1
-      - distro==1.9.0
-      - docutils==0.21.2
-      - fonttools==4.56.0
-      - kiwisolver==1.4.8
-      - matplotlib==3.10.1
-      - numpy==2.2.3
-      - packaging==24.2
-      - pillow==11.1.0
-      - pyparsing==3.2.1
-      - python-dateutil==2.9.0.post0
-      - pyyaml==6.0.2
-      - rospkg==1.6.0
-      - six==1.17.0
-prefix: /home/radtop/ros1/.conda/envs/mpc
